requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cssselect==1.1.0
dataclasses==0.8
elasticsearch==7.15.2
fake-useragent==0.1.11
flake8==4.0.1
frozenlist==1.2.0
geographiclib==1.52
geopy==2.2.0
googletransx==2.4.2
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.2.0
lxml==4.6.4
mccabe==0.6.1
multidict==5.2.0
numpy==1.19.5
oauthlib==3.1.1
pandas==1.1.5
parse==1.19.0
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
pyee==8.2.2
pyflakes==2.4.0
pyppeteer==0.2.6
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-socks==2.0.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-html==0.10.0
requests-oauthlib==1.3.0
schedule==1.1.0
six==1.16.0
soupsieve==2.3.1
tqdm==4.62.3
tweepy==4.4.0
tweet-parser==1.13.2
twint==2.1.21
typing-extensions==4.0.1
urllib3==1.26.7
w3lib==1.22.0
websockets==9.1
yarl==1.7.2
zipp==3.6.0
